requirements.txt
langchain
langchain-community
langchain-groq
faiss-cpu
python-dotenv